+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## 1.6.0-SNAPSHOT
 
+* Deprecated ServiceOption.ON_DEMAND_LOAD. While the option still exists
+  it has no effect. Applications are discouraged from using it.
+  All indexed services are now eligible for on-demand stop
+  (under memory pressure) and on-demand start. The usual rules still
+  apply; specifically, a service needs to be inactive in order to be
+  eligible for stop. Inactivity threshold is measured as a factor of the
+  maintenance interval. The stop delay factor is a host setting
+  and can be overridden via ServiceHost.setServiceStopDelayFactor().
+
+ * Added support to exclude built-in fields like 'documentVersion' when outputting to JSON.
+   Revamped JSON output methods to use JsonOptions enum instead of various boolean flags.
+   Deprecated JSON output methods that use boolean flags in favor of the new design.
+
 ## 1.5.7
 
 * Fix auth check for non-persisted stateful service on document-index GET.
@@ -23,11 +36,14 @@
     "kind": "com:vmware:xenon:services:common:SynchronizationRequest"
   }'
 
+
 ## 1.5.6
 
 * Upgrade Netty to [4.1.15](http://netty.io/news/2017/08/25/4-0-51-Final-4-1-15-Final.html) and
   netty-tcnative to [2.0.6](https://github.com/netty/netty-tcnative/releases/tag/netty-tcnative-parent-2.0.6.Final)
 
+
+* Instances java.nio.file.Path are now serializable to JSON
 
 ## 1.5.5